backend/requirements.txt
fastapi
uvicorn
pydantic
supabase
openai
langchain
python-dotenv
